ROFO-208 refactor: titleTopbar 컴포넌트화
